requirements.txt
selenium
requests
webdriver-manager
pytest
mysql-connector-python
allure-pytest